+++ analysis/import/SourceCodeParser/src/main/kotlin/de/maibornwolff/codecharta/importer/sourcecodeparser/SourceCodeParserMain.kt
@@ -3,6 +3,7 @@ package de.maibornwolff.codecharta.importer.sourcecodeparser
 import de.maibornwolff.codecharta.importer.sourcecodeparser.metricwriters.CSVMetricWriter
 import de.maibornwolff.codecharta.importer.sourcecodeparser.metricwriters.JSONMetricWriter
 import de.maibornwolff.codecharta.importer.sourcecodeparser.metricwriters.MetricWriter
+import de.maibornwolff.codecharta.serialization.OutputFileHandler
 import de.maibornwolff.codecharta.serialization.ProjectDeserializer
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
@@ -87,20 +88,21 @@ class SourceCodeParserMain(
     }
 
     private fun getMetricWriter(): MetricWriter {
-        val filePath = outputFile?.absolutePath ?: "notSpecified"
         return when (outputFormat) {
-            OutputFormat.JSON -> JSONMetricWriter(getOutputWriter(), filePath, compress)
-            OutputFormat.TABLE -> CSVMetricWriter(getOutputWriter())
+            OutputFormat.JSON -> JSONMetricWriter(getJsonOutputStream(), compress && outputFile != null)
+            OutputFormat.TABLE -> CSVMetricWriter(getCsvOutputWriter())
         }
     }
 
-    private fun getOutputWriter(): Writer {
+    private fun getCsvOutputWriter(): Writer {
         return if (outputFile == null) {
             OutputStreamWriter(outputStream)
         } else {
             BufferedWriter(FileWriter(outputFile!!))
         }
     }
+
+    private fun getJsonOutputStream() = OutputFileHandler.stream(outputFile?.absolutePath, outputStream, compress)
 
     companion object {
         @JvmStatic