+++ visualization/app/app.module.ts
@@ -1,3 +1,4 @@
+import "@angular/compiler" // todo this is needed for JIT compiler for ngColor within a downgraded component. Latest after full migration we can likely remove it again
 import "./app" // load AngularJS app first
 import "zone.js" // needs to be loaded before "@angular/core"
 import { APP_INITIALIZER, Inject, NgModule } from "@angular/core"
@@ -18,11 +19,11 @@ import { UnfocusNodesOnLoadingMapEffect } from "./codeCharta/state/effects/unfoc
 import { AddBlacklistItemsIfNotResultsInEmptyMapEffect } from "./codeCharta/state/effects/addBlacklistItemsIfNotResultsInEmptyMap/addBlacklistItemsIfNotResultsInEmptyMap.effect"
 import { dialogs } from "./codeCharta/ui/dialogs/dialogs"
 import {
-	threeSceneServiceProvider,
 	codeChartaServiceProvider,
-	threeOrbitControlsServiceProvider,
 	threeCameraServiceProvider,
-	threeRendererServiceProvider
+	threeOrbitControlsServiceProvider,
+	threeRendererServiceProvider,
+	threeSceneServiceProvider
 } from "./codeCharta/services/ajs-upgraded-providers"
 import { NodeContextMenuCardModule } from "./codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.module"
 import { OpenNodeContextMenuEffect } from "./codeCharta/state/effects/nodeContextMenu/openNodeContextMenu.effect"
@@ -33,7 +34,6 @@ import { LoadingFileProgressSpinnerComponent } from "./codeCharta/ui/loadingFile
 import { LoadingMapProgressSpinnerModule } from "./codeCharta/ui/toolBar/loadingMapProgressSpinner/loadingMapProgressSpinner.module"
 import { LoadingMapProgressSpinnerComponent } from "./codeCharta/ui/toolBar/loadingMapProgressSpinner/loadingMapProgressSpinner.component"
 import { BlacklistSearchPatternEffect } from "./codeCharta/ui/searchPanel/searchBar/blacklistSearchPattern.effect"
-import { EdgeMetricToggleComponent } from "./codeCharta/ui/edgeSettingsPanel/edgeMetricToggle/edgeMetricToggle.component"
 import { SearchPanelComponent } from "./codeCharta/ui/searchPanel/searchPanel.component"
 import { SearchPanelModule } from "./codeCharta/ui/searchPanel/searchPanel.module"
 import { UploadFilesButtonModule } from "./codeCharta/ui/toolBar/uploadFilesButton/uploadFilesButton.module"
@@ -58,6 +58,7 @@ import { ActionIconModule } from "./codeCharta/ui/actionIcon/actionIcon.module"
 import { ColorSettingsPanelModule } from "./codeCharta/ui/ribbonBar/colorSettingsPanel/colorSettingsPanel.module"
 import { ScreenshotButtonModule } from "./codeCharta/ui/screenshotButton/screenshotButton.module"
 import { SplitStateActionsEffect } from "./codeCharta/state/effects/splitStateActionsEffect/splitStateActions.effect"
+import { CopyToClipboardButtonModule } from "./codeCharta/ui/copyToClipboardButton/copyToClipboardButton.module"
 import { DownloadButtonModule } from "./codeCharta/ui/toolBar/downloadButton/downloadButton.module"
 
 @NgModule({
@@ -102,7 +103,8 @@ import { DownloadButtonModule } from "./codeCharta/ui/toolBar/downloadButton/dow
 		ColorSettingsPanelModule,
 		ScreenshotButtonModule,
 		DownloadButtonModule,
-		UploadFilesButtonModule
+		UploadFilesButtonModule,
+		CopyToClipboardButtonModule
 	],
 	providers: [
 		threeSceneServiceProvider,
@@ -119,7 +121,7 @@ import { DownloadButtonModule } from "./codeCharta/ui/toolBar/downloadButton/dow
 			multi: true
 		}
 	],
-	declarations: [EdgeMetricToggleComponent, ...dialogs],
+	declarations: [...dialogs],
 	entryComponents: [
 		AttributeSideBarComponent,
 		Export3DMapButtonComponent,
@@ -127,13 +129,13 @@ import { DownloadButtonModule } from "./codeCharta/ui/toolBar/downloadButton/dow
 		FocusButtonsComponent,
 		LoadingFileProgressSpinnerComponent,
 		LoadingMapProgressSpinnerComponent,
-		EdgeMetricToggleComponent,
 		SearchPanelComponent,
 		...dialogs
 	]
 })
 export class AppModule {
 	constructor(@Inject(UpgradeModule) private upgrade: UpgradeModule) {}
+
 	ngDoBootstrap() {
 		this.upgrade.bootstrap(document.body, ["app"], { strictDi: true })
 	}